requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-limiter==0.1.6
greenlet==3.2.1
h11==0.14.0
httpie==3.2.4
httptools==0.6.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
passlib==1.7.4
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
qrcode==8.1
redis==6.0.0b2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
